requirements.txt
PyMuPDF>=1.23.8
python-docx>=1.1.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
